Uploaded cleaned Color Test
I decided to clean up, and tune the color sensor code. This is the end result of that process.

Additionally, raw data recorded from the sensor has been recorded in a data sheet commented into the code. All tests were done at 20% frequency scaling, and is noted in the comment.

The most important change done in cleanup has been the removal of hard coded variables being kept in code, and has been taken out and placed into defines at the top of the file.